Firehose commit e646f57873bd86fc63d8624e44fc46c5a5c5d2a7 code generation
bump ASPN-LCM version

+++ lcm/c/include/aspn23_lcm_type_image_feature.h
@@ -74,9 +74,9 @@ struct _aspn23_lcm_type_image_feature
     /**
      * Description: Feature descriptor.
      * Units: none
-     * LCM Type: int16_t[descriptor_size]
+     * LCM Type: byte[descriptor_size]
      */
-    int16_t    *descriptor;
+    uint8_t    *descriptor;
 };
 
 /**
